Added local exceptions for EmployeeServiceImpl

+++ src/main/java/com/example/classicmodlesslaes/service/EmployeeServiceImpl.java
@@ -2,6 +2,8 @@ package com.example.classicmodlesslaes.service;
 
 import com.example.classicmodlesslaes.model.Employee;
 import com.example.classicmodlesslaes.repository.interfaces.EmployeeRepository;
+import com.example.classicmodlesslaes.service.exceptions.DataAccessException;
+import com.example.classicmodlesslaes.service.exceptions.EntityNotFoundException;
 import com.example.classicmodlesslaes.service.interfaces.EmployeeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,47 +22,88 @@ public class EmployeeServiceImpl implements EmployeeService {
 
     @Override
     public Employee getEmployeeById(int id) {
-        return employeeRepository.getEmployeeById(id);
+        Employee employee = employeeRepository.getEmployeeById(id);
+        if(employee == null){
+            throw new EntityNotFoundException("Employee with ID: " + id + " not found.");
+        }
+        return employee;
     }
 
     @Override
     public Employee saveEmployee(Employee employee) {
-        return employeeRepository.saveEmployee(employee);
+        try {
+            return employeeRepository.saveEmployee(employee);
+        }catch (Exception e){
+            throw new DataAccessException("Error saving employee.", e);
+        }
     }
 
     @Override
     public Employee updateEmployee(Employee employee) {
-        return employeeRepository.updateEmployee(employee);
+        if(!existingEmployee(employee.getEmployeeNumber())){
+            throw new EntityNotFoundException("Cannot update. Employee with ID: " + employee.getEmployeeNumber() + " not found.");
+        }
+        try {
+            return employeeRepository.updateEmployee(employee);
+        }catch (Exception e){
+            throw new DataAccessException("Error updating employee.", e);
+        }
     }
 
     @Override
     public int deleteEmployee(int id) {
+        if(!existingEmployee(id)){
+            throw new EntityNotFoundException("Cannot delte. Employee with ID: " + id + " not found");
+        }
         return employeeRepository.deleteEmployee(id);
     }
 
     @Override
     public List<Employee> findEmployeesByJobTitle(String jobTitle) {
-        return employeeRepository.findEmployeesByJobTitle(jobTitle);
+        List<Employee> employees = employeeRepository.findEmployeesByJobTitle(jobTitle);
+        if(employees == null || employees.isEmpty()){
+            throw new EntityNotFoundException("No employees found with the job title: " + jobTitle);
+        }
+        return employees;
     }
 
     @Override
     public List<Employee> findEmployeesWithoutSupervisors() {
-        return employeeRepository.findEmployeesWithoutSupervisors();
+        List<Employee> employees = employeeRepository.findEmployeesWithoutSupervisors();
+        if(employees == null || employees.isEmpty()){
+            throw new EntityNotFoundException("No employees from with supervisor");
+        }
+        return employees;
     }
 
     @Override
     public List<Employee> findByOfficeCode(String officeCode) {
-        return employeeRepository.findByOfficeCode(officeCode);
+        List<Employee> employees = employeeRepository.findByOfficeCode(officeCode);
+        if(employees == null || employees.isEmpty()){
+            throw new EntityNotFoundException("No employee found for office code: " + officeCode);
+        }
+        return employees;
     }
 
     @Override
     public Employee findSupervisorOfEmployee(int employeeId) {
-        return employeeRepository.findSupervisorOfEmployee(employeeId);
+        Employee supervisor = employeeRepository.findSupervisorOfEmployee(employeeId);
+        if(supervisor == null){
+            throw new EntityNotFoundException("No supervisor found for employee: " + employeeId);
+        }
+        return supervisor;
     }
 
     @Override
     public List<Employee> findEmployeesWithNoEmail() {
-        return employeeRepository.findEmployeesWithNoEmail();
+        List<Employee> employees = employeeRepository.findEmployeesWithNoEmail();
+        if(employees == null || employees.isEmpty()){
+            throw new EntityNotFoundException("No employees found without an email address");
+        }
+        return employees;
     }
 
+    private boolean existingEmployee(int id){
+        return employeeRepository.getEmployeeById(id) != null;
+    }
 }
